Initial menu system and asyncronous functions
A pretty simple framework in C has been created to load the correct
modules and start everything up. This will later handle things like TCP
connections and more C modules will be written to do some heavy
algorithms and compression.

All curses code is located in scheme so that it should, in theory, be
possible to replace all the UI code without even needing to recompile
the C code.

In scheme there is a basic menu system implemented, so far only the
single option menu works but a the same code will be utilized when doing
a multiple option menu. There is also the (util reader) module that
implements a system for asyncronous function calls when the interface is
waiting for input. This is only supposed to be used for simple UI stuff
such as removing old flash-messages and maybe simpler animations.
Everything heavier than that should be running in it's own thread.