Added module for printing colored text.
When passing a string to color-string in (util colors) a list of complex
characters is returned with attributes set according to the format of
the string.

^B - Makes the following text bold
^X - Inverts the following text
^Cnn - Sets the following text to the color pair nn
^U - Makes the following text underlined

The same control character can also be used to switch off the attribute
so for a string like "foo ^Bbar^B baz" only bar would be bold. To reset
color to default you use a plain ^C without any numbers afterwards.
Important to remember with ^C is that any '\' following the optional
numbers are ignored and not included in the string so to make only 5
colored in the string you can write "1234^C2\5^C\6789".