Changed color-string formating and fixed async bug
Changed the format that color-string uses so that it doesn't use control
characters but %-sequences instead. The new syntax is:

 - %b for bold text
 - %cnn for color using the color-pair nn, can optionally be terminated
   by ; if the following text is unknown and could be numbers.
 - %f for blinking (flashing) text.
 - %u for underlined text
 - %d for dim text, this does not work on the terminals I've tried.
 - %i for inverted text

There was also a bug in the async management that caused the whole
process to crash if the async threw an exception, this has been fixed so
that all exception thrown from asyncs should be dismissed.

A separate color-screen has been added that can be accessed from the
main-menu at slot c, this shows as many defined colors as possible.